fix: don't update twice
since we already make sure the state is not leaving we can assume the collator is
active. We only need to be careful when we add new states to the enum.
when there is a new state (e.g. inactive) this logic would brake.
Therefore we keep the `is_active` guard