feat: unlimited candidates (#244)
* step 1: count candidates

* refactor: rename state -> pool & pool -> top

* refactor: simple set_max_candidate_stake

don't adjust existing collator stake. If root wants to adjust the stake,
they need to adjust the existing stake manually with root calls.

* feat: try_insert_replace can insert without replace

if space in the set is left

* use safe math

* shouldn't be that hard

* refactor: remove unused methods

binary search is potentially dangerous becuase it might not find the correct item

* step2: allow joining when top candidates are full

* chore: 🧹

* docs: update some doc

* doc: OrderedSet::mutate

* feat: event for entering and leaving top candidates

* fix: OrderedSet edge case max size = 0 or 1

* doc: explain

* style: turbo fish ::<>

* fix: don't update twice

since we already make sure the state is not leaving we can assume the collator is
active. We only need to be careful when we add new states to the enum.
when there is a new state (e.g. inactive) this logic would brake.
Therefore we keep the `is_active` guard

* refactor: it's a candidate

* fix: make sure to always increase the stake

if the delegators set wasn't full we might not increase the stake.
better safe than sorry

* doc: imrpove comment

* refactor: rename limits to max/min Top Candidates

* refactor: rename to MinCollators

* fix: test

Co-authored-by: William Freudenberger <william@kilt.io>