split deposit owner remove in separate tx
Having everything in a single extrinsic is bad because it's not clear
who wants to remove the attestation.
In case an unauthorised subject wants to remove an attestation and the
tx gets send by the deposit owner, the attestation gets removed even if it should
be an unauthorized call. The owner didn't intend to remove the
attestation.