chore: update Polkadot dependencies to 0.9.39 (#500)
Fixes https://github.com/KILTprotocol/kilt-node/issues/495.

Since the `wasmtime` dependency has been bumped from 1.x to 6.x, the old
nightly was not good anymore since some inline stuff the new crate is
using was still not stable back then. Hence, I bumped the nightly to a
more recent version. Nevertheless, it cannot be TOO recent because of
[this issue](https://github.com/paritytech/substrate/issues/13636)
(which maybe has been fixed in 0.9.40 @weichweich?).

Anyway, updating to the new toolchain version added a whole bunch of
Clippy warnings which I also addressed in this PR, among which there was
one about the wrong declaration of the `parity-scale-codec` package,
which I have now used with its original name everywhere, instead of
aliasing it to `codec`.

---------

Co-authored-by: Adel Golghalyani <48685760+Ad96el@users.noreply.github.com>