feat: configurable origin check for 'commit_identity' (#526)
Identity information should be entirely under the control of the DID
subject. Until now, it was not the case as anyone could have called the
`commit_identity` extrinsic for any DID subject.
This PR changes the origin check for the `commit_identity` extrinsic of
provider pallet from `EnsureSigned` to a configurable origin check.
The XCM integration tests have also been updated to use the
`EitherOfDiverse<EnsureRoot<AccountId>, EnsureDidOrigin<DidIdentifier,
AccountId>>;` implementation.
Now, except for cases where democracy decides to do this operation, the
DID subject has to agree to it by signing the operation, which will then
be dispatched via the `submit_did_call` extrinsic of the DID pallet.

The feature that any account can pay for the fees still applies, as that
is the same also for DID-signed extrinsics.