feat: restrict charset of a DID service ID and endpoint (#535)
Fixes https://github.com/KILTprotocol/ticket/issues/2626.

Right now, we only check for `is_ascii()` for all elements of a DID
service. The [DID core spec](https://www.w3.org/TR/did-core/#services)
defines a service ID and a service endpoints as URIs by requiring the
presence of a `#`, which makes an ID and a service endpoint without the
leading `kilt:did:...` prefix a fragment by all means. This PR
implements this requirement, by only allowing fragment-allowed
characters as part of a service ID and endpoint, further restricting the
set to not include percent-encoded characters, for the sake of ease of
implementation, without deviating from the standard (which we now kinda
do). This requirement can be relaxed in the future, if needed.