fix: origin checks for provider and consumer pallets (#581)
Ensuring the origin is a DID origin for the implementor of the provider
pallet might not be enough. They need to check that the origin indeed
matches the identifier being acted upon. Replacing `EnsureOrigin` with
`EnsureOriginWithArg` gives access to that information.

Tests will ALL be implemented in a different PR 😄