feat: update templates and Dockerfile to support Zombienet (#587)
Zombienet uses the empty chainspec for a parachain, and that was not
supported by the node templates. Also, the Dockerfile assumes the
`kilt-parachain` is being run, but this is different now, and could also
be different in the future, so the binary target should be parametrized:
anyway, it is not necessary to compile the whole workspace if only a
specific binary is required.