feat: benchmarks (#585)
## fixes
[KILTProtocol/ticket#2551](https://github.com/KILTprotocol/ticket/issues/2551)

This pull request introduces benchmark logic for the DIP templates. To
calculate default weights, a dummy implementation of the pallets in the
Peregrine runtime is defined. After the PR is approved, I will remove
the dummy implementation.


## Leftover tasks

- [x] Make the provider and consumer binaries compiles for benchmarking.
Currently the `VersionedSiblingParachainDipStateProof` type does not
implement the `WorstCase` trait, which is a blocker to use it in
benchmarks
- [x] Adjust the `DipProofVerifier` logic to allow any calls to be
dispatched when being benchmarked
- [x] Make sure everything compiles
- [x] Make sure benchmarks can actually be run for both provider and
consumer runtimes

---------

Co-authored-by: Antonio Antonino <antonio@kilt.io>