fix: benchmarks (#598)
There was a circular call graph issue, where the deposit removal called
the hook, which in turn tried to remove the deposit again. A new
function on the `pallet-deposit-storage` is added which allows to delete
a deposit without calling into the hook.
Benchmarks now run fine for both `pallet-deposit-storage` and
`pallet-dip-provider` for both peregrine and DIP Provider template.