chore: add unit tests for DIP components (#609)
Partially fixes https://github.com/KILTprotocol/ticket/issues/2562.

## Tests for the DIP provider components

The DIP provider components are tested in this PR. Specifically:

* The `pallet-deposit-storage` pallet main logic.
* The `DepositHooks` hook for deleting an identity commitment when a
deposit is released directly. It used to leave in the Peregrine runtime
but it now lives in `runtime-common` and is generic over the runtime.
This was required for testing the hook logic.
* The `pallet-dip-provider` pallet main logic.
* The `FixedDepositCollectorViaDepositsPallet` hook for reserving and
releasing deposits when identity commitments are deleted. It uses a
different mock than the pallet mock.
* The `generate_proof` and `generate_commitment` functions, along with
the `DidMerkleRootGenerator`.
* The `LinkedDidInfoProvider`.

The PR also contains minor refactoring that was necessary to do proper
mocking of the different pieces of the runtime to test the components
above. The main changes are:

* Move from a multi-module to a single-module approach, where files do
not contain more than a module.
* Move the `DepositHooks` from being Peregrine-specific to being
runtime-agnostic, in `runtime-common`.

**This PR only adds tests for the DIP provider components. The DIP
consumer will be tested in a separate PR.**