fix: change DIP components storage hashers (#613)
Changing the storage hasher before we deploy on production, otherwise we
would then need a migration strategy. I think
we can disregard any migrations on our testnets, as the effort is not
justified.

I also lowered the limits of the DIP provider template, which were
higher than Peregrine, so that it is faster to generate the benchmark
worst case.

Related to https://github.com/KILTprotocol/kilt-node/pull/601, based on
top of https://github.com/KILTprotocol/kilt-node/pull/612.