fix: outdated integration tests (#632)
Fixes https://github.com/KILTprotocol/ticket/issues/3344.

Integration tests are outdated. Since KILT is now listed on HydraDX, any
tokens sent to the Omnipool account is automatically transferred to the
treasury, which makes the test cases fail. The integration test logic
must be updated to test the following:

1. Send KILT tokens from a KILT account to **its own** sovereign account
on HydraDX (implementation started in this PR for the V2 tests)
2. Send KILT tokens from HydraDX back to KILT (to implement in this PR)

---------

Co-authored-by: Adel Golghalyani <ad96el@gmail.com>