bench: add switch pallet benchmarks (#671)
Fixes https://github.com/KILTprotocol/ticket/issues/3511.

Adds benchmars to the asset-switch pallet. The benchmark logic could
definitely be improved and be made more generic, but I think for the
time being this is a good compromise. If we see that other projects are
struggling to run the benchmarks with their runtime, we can think about
how to improve them.