test: asset switch XCM components unit tests (#675)
Fixes https://github.com/KILTprotocol/ticket/issues/3514.

It adds extensive unit tests for the different XCM components. Because
each components has slightly different needs, and because unit tests are
supposed to be isolated from each other, I created different mocks. This
means that the different components and their test environments can
evolve at different speeds and independently.

The important bit is that most of the unit tests also run the
crate-level `try-runtime` logic, so we can be almost always sure that
the invariant we set out in the beginning holds also at the end of each
individual test case. The only places where the integrity tests are not
run, is where the component does not interact with the currency
sub-system at all, which has been mocked as well. All other cases always
run integrity tests.