chore: CI fixes (#676)
Fixes https://github.com/KILTprotocol/ticket/issues/3495, based on top
of https://github.com/KILTprotocol/kilt-node/pull/675.

I had to remove the tests for the `RuntimeCall` size, since the newly
introduced `pallet_assets` has a size of 896 bytes, which is huge. I
noticed that none of the chains where the pallet is deployed and uses
`MultiLocation` as asset ID has this test anymore... Maybe that will
change in the future again.
Also, the `try-runtime` tests fail the idempotency tests for the same
`pallet_assets` pallet. This should be solved by the
`VersionedMigration` struct, but in 1.1 (the version this PR is based
on) it was still marked as experimental. We might be able to solve that
once we rebase on top of 1.7.