feat: switch functionality (#658)
Fixes https://github.com/KILTprotocol/ticket/issues/3467.

Not meant to be merged until all other stacked PRs will be merged into
this.

Adds the first round of features to the swap pallet, specifically it
includes a swap pool that supports only the runtime local currency,
swappable 1:1 (not configurable yet) with a configurable remote token.
The XCM fees are paid in a local fungible token as injected by the
runtime. The pallet also exposes two implementations of the `IsReserve`
trait that must be added to the runtime to allow the configured tokens
(to swap back to local currency and to pay for XCM fees) to be
reserve-transferred to this chain. That code is not yet part of this PR.

Additional TODOs for the pallet is both within the code where
applicable, and in the pallet `README.md` file.

---------

Co-authored-by: Adel Golghalyani <48685760+Ad96el@users.noreply.github.com>
Co-authored-by: Adel Golghalyani <ad96el@gmail.com>