chore: add peregrine-dev id (#715)
Because we match a Peregrine chainspec if the name itself has
`peregrine` in it, currently we cannot spin up a Zombienet-based network
using Peregrine dev because its name is only `dev`, which would fail to
match. I added a second way to generate a peregrine dev chainspec, which
is `peregrine-dev`, which contains the word `peregrine` in it, and would
allow us to spin up a Zombienet network.