chore: add events for deposit owner changes (#738)
Fixes https://github.com/KILTprotocol/ticket/issues/3074.

Cc @kilted-andres 

I tried to keep the event structure in all pallets the same, with the
same name and the same fields `id`, `from` and `to`. The `id` field has
a different type depending on the pallet, but the `from` and `to` are
all the same.