feat: create pool (#768)
##  Fixes: [#3657](https://github.com/KILTprotocol/ticket/issues/3657)

This PR introduces the feature to create new bonded pools.

For each newly created currency, the metadata are updated, and the team
is reset. The currently implemented trait will be removed once the
Polkadot version provides it.

The generated pool_id is also used as a proxy account where the
collateral is accumulated.

The deposit system may change in the future, but a basic implementation
is provided for now.