feat: support for unique linking (#782)
Part of https://github.com/KILTprotocol/ticket/issues/3650. Built on top
of https://github.com/KILTprotocol/kilt-node/pull/781.

## Trade-off

I chose to go this way instead of providing an optional counter, because
providing a counter would require one of the following two approaches:

1. Transform the storage double map into a counted one, requiring a
migration also for our currently-deployed pallet, which I wanted to
avoid
2. Not use a counter, but iterate every time to make sure there are
still "spots" left for the current DID. This would require changing the
benchmarking logic as now we have a potentially unbounded iteration
happening. I also wanted to avoid that.

Hence, the solution was to provide a somehow more limited feature of
simply specifying whether the links are expected to be unique per DID or
not. This, as long as we set `false` for our deployed pallets would not
require any storage migration, and does not require any changes in the
benchmarks, so I found it a good compromise.