fix: runtime API implementations not exported (#809)
Fixes https://github.com/KILTprotocol/ticket/issues/3688.

Due to how the `impl_runtime_apis` macro works, it is expected to be
included in a runtime's `lib.rs` file. Since we split up the runtimes,
the runtime APIs were not included anymore. I found a related issue in
the subxt repo: https://github.com/paritytech/subxt/issues/1873.

Because the trait definition generated by the macro is local, we can't
import it in the `lib.rs` module for the `construct_runtime` macro to
pick up the right implementation, so we need a workaround that basically
introduces a new marker trait, and we import _that_ in the `lib.rs` file
so that the right implementation of `Runtime::runtime_metadata()` is
picked up. More details about the issue are presented in the subxt
ticket.

## How to test

Spin up a chopsticks deployment after building peregrine and spiritnet
runtime, and verify that the runtime APIs are there now.