feat: implement `MutateHold` for `pallet-deposit-storage` (#823)
Fixes https://github.com/KILTprotocol/ticket/issues/3702.

This PR implements the `MutateHold` trait for the deposit storage pallet
by introducing a `PalletDepositStorageReason` type that, in the runtime,
is required to implement the runtime's `RuntimeHoldReason`.

Because the deposit pallet assumes that the deposit payer can reclaim
the deposit back, we can't add the `MutateHold` implementation without
either 1. requiring a migration to distinguish which deposits are
reclaimable and which are not, or 2. introducing a new storage map that
only keeps track of "system" deposits, aka deposits that are part of
some runtime logic and that are only supposed to be changed by
interacting with the `MutateHold` trait and not directly via the pallet
exposed extrinsic(s).

## How to use it in the bonding curves

When switching to the deposit pallet as the deposit fungible, the
bonding curve pallet has to define a namespace, and a unique key per
deposit to be taken. For instance, the tuple (asset ID, account ID)
could be such a unique key. Example of integration PR:
https://github.com/KILTprotocol/kilt-node/pull/827. I can provide
support for integrating this into the runtime once the runtimes have
been updated to have the required pallets.