feat: add support for metadata hash checks (#835)
We introduce a new feature for peregrine and spiritnet runtimes, which
adds the metadata hash signed extension, for wallets that need it.

The Gitlab build pipeline has also been updated to include the new
feature only when building the production WASMs.

## How to test

Change any integration tests in the SDK based on this PR
(https://github.com/KILTprotocol/sdk-js/pull/924), and verify the
metadata is verified (e.g., by enabling logs via a Chopsticks
deployment).