feat: prevent DID deletions if there's linked resources (#843)
Fixes https://github.com/KILTprotocol/ticket/issues/3673.

I defined a new collection of hooks for the DID pallet, which can be
extended with new functions if we need to check conditions for some
other DID operations. I provide a bunch of implementations to check for
related resources.

~The only limitation to the current solution is that we don't have a
precise worst case weight calculation for the hook, since we would need
to benchmark the storage accesses, and the pallets use the old
benchmarking logic which is not very flexible. I will work on the
benchmarks and on updating the actual weights in a separate PR.~
Apparently, [it's enough to use the expected `MaxEncodedLen` as the
proof
size](https://substrate.stackexchange.com/questions/11842/calculate-proof-size-of-a-function-that-reads-state/11843#11843),
so I switched to that in the [final
commit](https://github.com/KILTprotocol/kilt-node/pull/843/commits/72fc198d24b84e19b7d04a1da0134742f0f2b9d4).

## How to test

Spin up a Chopsticks instance, claim a DID, claim a web3name, try to
delete the DID: it will fail. Delete the web3name. Try to delete the
DID: it will succeed.
The same can be tried with linked accounts.